refactor: remove duplicate background and popup implementations, keeping simpler versions
